Added NSURLConnection Extension
The current registration of protocols let SuperMock record  any call
using NSURRequest and NSURLSession init with
NSURLSession.sharedSession().
The creation of custom sessions with NSURLSession(configuration:
NSURLSessionConfiguration) do not have the protocols registered.
To sort this out as soon as we create a NSURLSessionConfiguration we
need to call the method addProtocols, in this way the protocols get
registered and Supermock work perfectly.

A preferred solution was override the method
defaultSessionConfiguration, but the override need to work in Obj-c as
well and this is currently hard to solve. I have tried a couple of way
but not arrived to a nice solution.

For now this is a working solution, but need to be specified in the
documentation that for creation of custom session let
NSURLSession(configuration: NSURSessionConfiguration)
need to be called configuration.addprotocols() for add the superMockUrl
protocols